+++ test/libsolidity/syntaxTests/constantEvaluator/type_reference.sol
@@ -1,4 +1,3 @@
 int[L] constant L = 6;
 // ----
 // TypeError 5462: (4-5): Invalid array length, expected integer literal or constant expression.
-// DeclarationError 9259: (0-21): Constants of non-value type not yet implemented.