+++ test/libsolidity/syntaxTests/constantEvaluator/type_reference_in_contract.sol
@@ -3,4 +3,3 @@ contract C {
 }
 // ----
 // TypeError 5462: (21-22): Invalid array length, expected integer literal or constant expression.
-// DeclarationError 9259: (17-38): Constants of non-value type not yet implemented.