+++ test/libsolidity/syntaxTests/iceRegressionTests/const_struct_with_mapping.sol
@@ -5,4 +5,4 @@ contract C {
     S public constant e = 0x1212121212121212121212121212121212121212;
 }
 // ----
-// DeclarationError 9259: (71-135): Constants of non-value type not yet implemented.
+// DeclarationError 3530: (71-135): The type contains a (nested) mapping and therefore cannot be a constant.