+++ test/libsolidity/syntaxTests/nameAndTypeResolution/105_constant_input_parameter.sol
@@ -3,4 +3,3 @@ contract test {
 }
 // ----
 // DeclarationError 1788: (31-55): The "constant" keyword can only be used for state variables or variables at file level.
-// DeclarationError 9259: (31-55): Constants of non-value type not yet implemented.